move mixin example
