+++ cypress/component/basic/mixins/README.md
@@ -1,0 +1,5 @@
+# mixins
+
+You can register global mixin component instances when mounting the a component. See [spec.js](spec.js)
+
+See [Vue global mixin docs](https://vuejs.org/v2/guide/mixins.html#Global-Mixin)
